+++ documentation/changelog.md
@@ -5,6 +5,7 @@ This is the changelog for `casual` and all changes are listed in this document.
 
 ### Fixes
 - tools: build-server does not generate function name via argument ([#568](https://github.com/casualcore/casual/issues/568))
+- common: transaction::global::ID istream asserts on hex <= 64B ([#570](https://github.com/casualcore/casual/issues/570))
 
 ## [1.7.7] - 2025-05-27